Refactoring folder structure to support auto adding package to xcode
